==> 01. data-retreive.ipynb <==
import numpy as np
import pandas as pd
import requests, zipfile, io, os
import json, glob, re, tmdb
from importlib import reload
from utils import create_directory, update_progress
import time

tmdb_api_key = "a67ce78b9ac0e0f7df8c2fc16c2fd30a"

zip_file_url = "http://files.grouplens.org/datasets/movielens/ml-100k.zip"

print("Downloading movielens data...")
r = requests.get(zip_file_url, stream=True)
z = zipfile.ZipFile(io.BytesIO(r.content))
z.extractall()
print("Done.")

users_cols = ['user_id', 'age', 'sex', 'occupation', 'zip_code']
users = pd.read_csv('ml-100k/u.user', sep='|', names=users_cols, encoding='latin-1')

users.shape

users.head()

genre_cols = [
    "genre_unknown", "Action", "Adventure", "Animation", "Children", "Comedy",
    "Crime", "Documentary", "Drama", "Fantasy", "FilmNoir", "Horror",
    "Musical", "Mystery", "Romance", "SciFi", "Thriller", "War", "Western"
]
items_cols = ['movie_id', 'title', 'release_date', "video_release_date", "imdb_url"] + genre_cols
items_raw = pd.read_csv('ml-100k/u.item', sep='|', names=items_cols, encoding='latin-1')

items_raw.shape

items_raw.head()

ratings_cols = ['user_id', 'movie_id', 'rating', 'unix_timestamp']
ratings = pd.read_csv('ml-100k/u.data', sep='\t', names=ratings_cols, encoding='latin-1')

ratings.shape

ratings.head()

create_directory("/data")

users.to_csv("data/users.csv", index=None)
items_raw.to_csv("data/items_raw.csv", index=None)
ratings.to_csv("data/ratings.csv", index=False)

links = pd.read_csv("data/links.csv", dtype={"tmdb_id": 'Int64' })
links.head()

links.shape

sample_id = links.tmdb_id[1]

movie_response = requests.get(f"https://api.themoviedb.org/3/movie/{sample_id}?api_key={tmdb_api_key}")
movie_response.text

tmdb_ids = links.tmdb_id.tolist()

def get_movie_features(tmdb_ids):
    num = len(tmdb_ids)
    tick = 0
    features = []
    for i in tmdb_ids:
        # feature = tmdb.movie(i)
        feature = requests.get(f"https://api.themoviedb.org/3/movie/{i}?api_key={tmdb_api_key}")

        if feature != False: 
            features.append(feature.text)
        
        tick = tick + 1
        update_progress(tick / num)
    
    update_progress(1)
    return features

def save_movie_features(features):
    outfile = "data/features.json"
    with open(outfile, 'w') as fout:
            json.dump(features, fout)

features = get_movie_features(tmdb_ids)
save_movie_features(features)

len(features)

text_features = []

for i in features:
  text_features.append(i.text)

text_features[0]

save_movie_features(text_features)
