==> 03. transform.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt


users = pd.read_csv("data/users.csv")
items = pd.read_csv("data/items.csv")
ratings = pd.read_csv("data/ratings.csv")

users.iloc[0]

items.iloc[0]

ratings.iloc[0]

# Preprocessing

from reco.preprocess import encode_user_item

encoded_ratings, user_encoder, item_encoder = encode_user_item(ratings, "user_id", "movie_id", "rating", "unix_timestamp")

ratings['rating'] = ratings['rating'].values.astype(np.float32)
min_rating = min(ratings['rating'])
max_rating = max(ratings['rating'])

# Data Splitting

from reco.preprocess import random_split, user_split, sample_data

sampledf = sample_data()

## Random split

train_random, val_random, test_random = random_split(sampledf, [0.6, 0.2, 0.2])

train_random

val_random

test_random

## User Chronological Split

train_chrono, val_chrono, test_chrono = user_split(sampledf, "timestamp", "user_index", [0.6, 0.2, 0.2])

train_chrono

val_chrono

test_chrono
