==> 05. feature-extractor.ipynb <==
import sys
sys.path.append("../")

import warnings
warnings.filterwarnings("ignore")

import numpy as np
import pandas as pd
import re

items_raw = pd.read_csv("data/items_raw.csv")
item_features = pd.read_csv("data/item_features.csv")

items_raw["release_date"] = pd.to_datetime(items_raw.release_date, infer_datetime_format=True)
items_raw["year"] = items_raw.release_date.apply(lambda x: str(x.year))

items_main = items_raw.drop(['video_release_date', 'release_date', 'imdb_url'], axis=1).copy()

items_addtl = item_features[['overview', 'original_language', 'runtime', 'vote_average', 'vote_count', "movie_id"]].copy()

items = pd.merge(left=items_main, right=items_addtl, on="movie_id", how="left")

items.head()

items.overview.isna().sum()

items.overview.fillna("None", inplace=True)

import spacy
nlp = spacy.load('en_core_web_lg')

doc = nlp(items["overview"][0])

doc.vector

def word_vec(sentence):
    doc = nlp(sentence)
    return doc.vector

%%time
overview_embedding = items["overview"].apply(word_vec)

overview_embedding = overview_embedding.to_list()

overview_embedding_list = []
for vec in overview_embedding:
    overview_embedding_list.append(vec.tolist())

len(overview_embedding_list)

overview_embedding_df = pd.DataFrame(overview_embedding_list)

overview_embedding_df.head()

items.columns

item_similarity_df = pd.concat([
    items[['movie_id', 'genre_unknown', 'Action', 'Adventure',
       'Animation', 'Children', 'Comedy', 'Crime', 'Documentary', 'Drama',
       'Fantasy', 'FilmNoir', 'Horror', 'Musical', 'Mystery', 'Romance',
       'SciFi', 'Thriller', 'War', 'Western']], 
    overview_embedding_df],
    axis=1)

item_similarity_df.head()

# Build nearest neighbor model

from reco.recommend import get_similar
from sklearn.neighbors import NearestNeighbors
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
import PIL

%%time
item_distances, item_similar_indices = get_similar(overview_embedding_df, 5)

item_similar_indices
