==> 11. Implicit MF copy.ipynb <==
# Matrix Factorisation - Implicit

import sys
sys.path.append("../")

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
from reco.preprocess import encode_user_item, random_split, neg_feedback_samples

%matplotlib inline

## Prepare the data

df_ratings = pd.read_csv("data/ratings.csv")
df_items = pd.read_csv("data/items.csv")
df_implicit = neg_feedback_samples(df_ratings, rating_threshold=3, ratio_neg_per_user=1)
df_implicit["unix_timestamp"] = 1


df_ratings.shape, df_implicit.shape

DATA, user_encoder, item_encoder = encode_user_item(df_implicit, "user_id", "movie_id", "rating", "unix_timestamp")

n_users = DATA.USER.nunique()
n_items = DATA.ITEM.nunique()
n_users, n_items

max_rating = DATA.RATING.max()
min_rating = DATA.RATING.min()
min_rating, max_rating

train, test = random_split(DATA, [0.8, 0.2])

train.shape, test.shape

# Implicit Matrix Factorisation

from keras.models import Model
from keras.layers import Input, Embedding, Flatten, Dot
from keras.regularizers import l2
from sklearn.neighbors import NearestNeighbors

def ImplicitMF (n_users, n_items, n_factors):
    
    # Item Layer
    item_input = Input(shape=[1], name='Item')
    item_embedding = Embedding(n_items, n_factors, 
                               embeddings_regularizer=l2(1e-6), 
                               name='ItemEmbedding')(item_input)
    item_vec = Flatten(name='FlattenItemsE')(item_embedding)

    # User Layer
    user_input = Input(shape=[1], name='User')
    user_embedding = Embedding(n_users, n_factors, 
                               embeddings_regularizer=l2(1e-6), 
                               name='UserEmbedding')(user_input)
    user_vec = Flatten(name='FlattenUsersE')(user_embedding)

    # Dot Product of Item and User
    rating = Dot(axes=1, name='DotProduct')([item_vec, user_vec])
    
    # Model Creation
    model = Model([user_input, item_input], rating)
    
    # Compile Model
    model.compile(loss='binary_crossentropy', optimizer="sgd")
    
    return model

n_factors = 40
model = ImplicitMF(n_users, n_items, n_factors)

%%time
output = model.fit([train.USER, train.ITEM], train.RATING, 
                                  batch_size=128, epochs=5, verbose=1, validation_split=0.2)

def get_embedding(model, name):
    embedding = model.get_layer(name=name).get_weights()[0]
    return embedding

item_embedding = get_embedding(model, "ItemEmbedding")
user_embedding = get_embedding(model, "UserEmbedding")

def get_similar(embedding, k):
    model_similar_items = NearestNeighbors(n_neighbors=k, algorithm="ball_tree").fit(
        embedding
    )
    distances, indices = model_similar_items.kneighbors(embedding)

    return distances, indices

%%time
item_distances, item_similar_indices = get_similar(item_embedding, 5)

item_similar_indices

def show_similar(item_index, item_similar_indices, item_encoder, df_items):
    s = item_similar_indices[item_index]
    movie_ids = item_encoder.inverse_transform(s)

    titles = []
    for movie_id in movie_ids:
        movie = df_items[df_items["movie_id"] == movie_id]
        if not movie.empty:
            # Extract just the string value from the Series
            title = movie['title'].values[0]
            titles.append(title)

    if not titles:
        print("No images found to display")
        return

    for title in titles:
        print(title)

show_similar(0, item_similar_indices, item_encoder, df_items=df_items)
